chatbot/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "chatbot"
version = "0.1.0"
readme = "README.md"
requires-python = "==3.13.7"

dependencies = [
    "fastapi==0.124.0",
    "httpx==0.28.1",
    "uvicorn[standard]==0.38.0",
    "pydantic==2.12.5",
    "pytz==2025.2",
    "pillow==12.0.0",
    "langfuse==3.10.5",
    "bs4==0.0.2",
    "jinja2==3.1.6",
    "opentelemetry-exporter-prometheus==0.60b0",
    "python-json-logger==4.0.0",
    "pydantic-ai==1.28.0",
    "aio-pika==9.5.8",
    "opentelemetry-instrumentation-aio-pika==0.60b0",
    "pyjwt==2.10.1",
    "unleashclient==6.4.1",
    "logfire[fastapi]==4.16.0",
    "sqlalchemy[asyncio]==2.0.44",
    "asyncpg==0.31.0",
    "alembic==1.17.2",
    "psycopg2-binary==2.9.11",
    "alembic-postgresql-enum==1.8.0",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.mypy]
files = ["src"]
strict = "true"
plugins = ['pydantic.mypy']
exclude = ["src/chatbot/openapi/generated"]
cache_fine_grained = "true"
namespace_packages = true

[tool.ruff]
line-length = 250
src = ["src"]
extend-exclude = ["src/chatbot/openapi/generated", "tests/test_ask_ai_api.py"]
lint.select = [
    # pycodestyle
    "E",
    # Pyflakes
    "F",
    # pyupgrade
    "UP",
    # flake8-bugbear
    "B",
    # isort
    "I",
]
lint.ignore = [
  "E501" # TODO: After open MR is merged
]
lint.flake8-bugbear.extend-immutable-calls = ["fastapi.Header"]

[dependency-groups]
dev = [
    "mypy==1.19.0",
    "ruff==0.14.8",
    "pytest==9.0.2",
    "pytest-asyncio==1.3.0",
    "testcontainers[postgres]==4.13.3",
    "inline-snapshot[dirty-equals]==0.31.1",
    # NOTE: https://github.com/koxudaxi/fastapi-code-generator/issues/495
    "click<8.2.0",
    "fastapi-code-generator==0.5.4",
    "dirty-equals==0.11",
]

[tool.uv.sources]
# NOTE: The fork includes Flip-specific fixes.
fastapi-code-generator = { git = "https://github.com/getflip/fastapi-code-generator.git" }
